chat/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
jinja2
python-jose
passlib[bcrypt]
uvicorn[standard]
